fix(client): share capability catalog uniqueness macro body and use feature-gated export for direct production test coverage
- Introduce unique-macro.inc.rs as single source for the assert body.
- Keep production assert private (no #[macro_export]) in openlark-client.
- Add internal '_test-catalog-unique' feature to export the production macro.
- Make openlark-capability-unique depend on client with the feature and re-export the macro.
- Update harness, ui tests, comments; bless updated .stderr expectations.
- This eliminates manual duplication and ensures compile-fail tests exercise the actual production implementation (per #423/#455 and Codex review).

Addresses latest Codex review findings on PR #455:
- P2: no more duplicated macro bodies.
- Medium: trybuild now covers production path directly.

All local tests (client + harness trybuild) pass; git diff --check clean.
