+++ crates/openlark-capability-unique/tests/ui.rs
@@ -1,8 +1,10 @@
 //! compile-fail：generation-time catalog uniqueness（#423 / #455）
 //!
-//! 此测试通过 openlark-capability-unique 导出的宏来验证错误格式。
-//! 实现体与 openlark-client 生产路径中的保持一致（手动同步）。
-//! 生产路径在每次编译 catalog 时执行断言。
+//! 此测试通过 openlark-capability-unique 导出的宏验证错误格式。
+//! 宏实现臂直接 include 自 openlark-client 的生产源码（unique-macro.inc.rs）。
+//! 因此测试覆盖的是与生产相同的逻辑实现。
+//!
+//! 生产路径在 client 每次生成 catalog 时也会执行该断言。
 
 #[test]
 fn capability_catalog_generation_time_uniqueness() {
