+++ crates/openlark-capability-unique/tests/ui/duplicate_capability_field.stderr
@@ -1,14 +1,10 @@
-error[E0428]: the name `auth` is defined multiple times
+error: cannot find macro `__capability_catalog_unique_body` in this scope
  --> tests/ui/duplicate_capability_field.rs:4:5
   |
 4 | /     openlark_capability_unique::assert_capability_catalog_unique! {
 5 | |         { field: auth, name: "auth" },
 6 | |         { field: auth, name: "auth" },
 7 | |     }
-  | |     ^
-  | |     |
-  | |_____`auth` redefined here
-  |       previous definition of the module `auth` here
+  | |_____^
   |
-  = note: `auth` must be defined only once in the type namespace of this module
   = note: this error originates in the macro `openlark_capability_unique::assert_capability_catalog_unique` (in Nightly builds, run with -Z macro-backtrace for more info)
